+++ src/main/java/com/clinica/model/User.java
@@ -9,7 +9,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -22,7 +22,7 @@ public class User {
     // Constructors
     public User() {}
 
-    public User(int id, String name, String email, String password,boolean isActive, boolean isAdmin) {
+    public User(Long id, String name, String email, String password,boolean isActive, boolean isAdmin) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -32,8 +32,8 @@ public class User {
     }
 
     // Getters & Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
